Refactor Export, Add tests
Refactor the data output format to use simple arrays for both the
required modules as well as the calls.

Having the variable names for the required modules is totally
nonsensical outside of the detection run.

The "true" keys of the former calls map are useful for easily removing
duplicates, but as final data format it's not optimal.

We keep a *Map data field around in both cases and transform the
requires map / nested calls map to flat arrays on Program.exit.

We add test cases for two use cases (direct module functions and module
member functions) and make sure that methods are properly detected
and ignored.