Refactor internal data-structures
Call order issues actually prevent our previous approach from working
outside of the tests :\

It could happen that a call is recorded while the dataStore was already
cleaned up for that module.

We change the way we deal with our storage in that we have a internal
format that we do not expose directly but copy/flatten into the previous
export format on Data.get().

Internally we just update the internal storage without cleaning up on
Program.exit. Cleanup still happens manually via Data.clear()