Add webpack plugin
Move the webpack plugin out my project into the plugin. Seems to work
fine.

It is not the only way to use the collected data but should work for
many cases and gives an idea how to do it general

Add documentation for the webpack plugin / a sourceRoot config example