Support ES6 imports
Add support for the following import constructs

    import tracked from "./service/tracked"

for both module functions directly on the default export

    tracked()

as well as member functions

    tracked.fn()

We also support something we did not support with ES5 because it is way
more natural in ES6

   import { trackedMember } from "./service/tracked"

   import { trackedMember as alias } from "./service/tracked"

"import *" variants are *not* supported so far.

Tests
-----

We restructure the tests to read the test modules to be tracked from
files in test/test-modules/.

Adds tests for above imports.