Fix path handling
With our new webpack multi-bundle configuration, the tracking of usages
wasn't working right because there was still a bug in the path handling
inside the trackVar function.

The test was also a bit oversimplified in that it would reuse always the
same module path for all its tests.

We make the test cases a bit stricter in that regard and add one extra
to check for support of sub directories to validate the fix for this bug.