Support multiple args, object literals (breaking)
Function tracking was reduced so far to track one String argument for a
function and potentially ignoring all other arguments as varArgs.

This commit breaks the data format we generate by evaluating all
arguments to a tracked function. That means that the named tracked
function records now contain an array of argument arrays each evaluated
statically.

Extends the static evaluation to support object expressions with
both literal and identifier keys and literal values. Any other kind of
object expression, either containing computed properties or non-literal
values is ignored as a whole.

Changes the varArgs configuration to support numeric values. A value of
`true` is equal to 1 and the behavior so far.

A varArgs value of `2` would capture the first two arguments as static
values and then ignore everything after the second argument as varArgs.