Allow empty args
Sometimes it might be useful to have a tracked function that is called
without any argument, just recording the fact that it is called
somewhere in the code. We should not stand in the way of that.

So we accept any arguments array as long as all non-varArg args evaluate
to something statically now.