Add typescript tests
I started the typescript endeavor after my initial testing outside of
the project in my typescript playground seemed to indicate it not
working.

Then I went through the elaborate effort to mirror the existing tests
here with typescript files (This time with the detected services
actually existing as skeleton definitions).

Then I added the typescript config changes to the new transform method
and added "@babel/preset-typescript" to our dev dependencies.

And then I wrote/repeated the existing Javascript tests with slightly
different paths, only to discover that it is indeed all working as
expected if we add the babel preset.

Very anticlimactic, but I guess not totally useless now that it all
exists.